(HTCONDOR-2742)  Extend the protocol to all the shadow to command the starter to run a diagnostic and return the result.

+++ src/condor_shadow.V6.1/NTreceivers.cpp
@@ -22,6 +22,7 @@
 #include "condor_classad.h"
 #include "condor_debug.h"
 #include "condor_io.h"
+#include "guidance.h"
 #include "pseudo_ops.h"
 #include "condor_sys.h"
 #include "baseshadow.h"
@@ -2345,7 +2346,7 @@ case CONDOR_getdir:
 
 		errno = 0;
 		ClassAd guidanceAd;
-		rval = pseudo_request_guidance(requestAd, guidanceAd);
+		rval = static_cast<int>(pseudo_request_guidance(requestAd, guidanceAd));
 		terrno = (condor_errno_t)errno;
 		dprintf( D_SYSCALLS, "\trval = %d, errno = %d\n", rval, terrno );
